moved fetching resources after add/edit/delete to sagas
