+++ src/redux/sagas/specificResourcesSaga.js
@@ -5,7 +5,7 @@ import axios from "axios";
 function* fetchSpecificResources(action) {
 
     try {
-        const response = yield axios.get(`/api/adminprevcare/specificresources/${action.payload}`);
+        const response = yield axios.get(`/api/adminprevcare/specificresources/${action.payload.categoryId}`);
 
         yield put({ 
             type: 'SET_SPECIFIC_RESOURCES',
@@ -18,30 +18,30 @@ function* fetchSpecificResources(action) {
 }
 
 function* addResource(action) {
-    console.log('in add resource');
-
+    
     try {
-        
+        const categoryId = action.payload.categoryId;
         const dataToSend = action.payload;
         console.log('data is', dataToSend);
 
         yield axios.post(`/api/adminprevcare/specificresources/${action.payload.categoryId}`,  dataToSend );
 
+        yield put({type: 'FETCH_SPECIFIC_RESOURCES' , payload: {categoryId: categoryId}});
     } catch (err) {
         console.error('Error adding specific resource', err);
     }
 }
 
 function* updateResource(action) {
-    console.log('in update resource');
-
+    
     try {
-        
+        const categoryId = action.payload.categoryId;
         const dataToSend = action.payload;
         console.log('data is', dataToSend);
 
         yield axios.put(`/api/adminprevcare/specificresources/${action.payload.categoryId}`,  dataToSend );
 
+        yield put({type: 'FETCH_SPECIFIC_RESOURCES' , payload: {categoryId: categoryId}});
     } catch (err) {
         console.error('Error updating specific resource', err);
     }
@@ -49,12 +49,15 @@ function* updateResource(action) {
 
 function* deleteResource(action) {
     try {
-      // console.log('in delete pin saga, pin is', action.payload);
-      yield axios.delete(`/api/adminprevcare/specificresources/${action.payload.categoryId}`, {data: action.payload});
-  
-    } catch (error) {
-      console.error('Error deleting resource:', error);
-      alert('could not delete resource');
+        const categoryId = action.payload.categoryId;
+
+        yield axios.delete(`/api/adminprevcare/specificresources/${action.payload.categoryId}`, {data: action.payload});
+            
+        yield put({type: 'FETCH_SPECIFIC_RESOURCES' , payload: {categoryId: categoryId}});
+
+        } catch (error) {
+        console.error('Error deleting resource:', error);
+        alert('could not delete resource');
     }
   }
